requirements.txt
torch
torchvision
segmentation-models-pytorch
pandas
pyyaml
Pillow

